feat: clippy -D warnings clean, wire SyncController+FramePool into playback loop, round 11
Clippy: workspace --all-targets --all-features -- -D warnings now passes (0 warnings, ~30 fixes, no public API changes). Playback loop: SyncController drop/duplicate by A/V drift (demux_worker) + FramePool scratch reuse for YUV->RGBA (native_pipeline). Round 11: RotateFilter (filter), yuv422p_frame_to_rgba (scale), apply_gain_f32/i16 (resample). Tests: player 80->85, filter 13->17, scale 14->16, resample 20->24. Also add .gitignore for .opencode/ and .codebase-memory/ runtime state.

+++ rsmpeg-filter/src/mirror.rs
@@ -20,8 +20,8 @@ impl MirrorFilter {
     /// swap but is preserved in value because the whole pixel is exchanged. For RGBA,
     /// `frame.data[0]` is a single tightly-packed plane with a 4-byte stride.
     pub fn apply(&self, frame: &mut Frame) {
-        let w = frame.width as usize;
-        let h = frame.height as usize;
+        let w = frame.width;
+        let h = frame.height;
         if w == 0 || h == 0 {
             return;
         }
